Patch from Tom Brown:
Fix for a few gpx parsing crashes

ZeroDivisionError happened in parse_gpx and parse_segment with special files
with no time delta.

I frequently get "Property total_meters must be a float" when uploading gpx
created by garmin. This change fixes it by filtering out short segments and
changing dist in parse_gpx to a float. This change tests that laps can be
inserted into the app engine store.

I had to make the test data longer to avoid "Average speed can not be higher
than maximum speed". The problem is that distance_list and speed_list are forced
to be the same length by inserting a 0 at the start of speed_list, so for 3
point segments total_distance / total_time is easily greater than
max(movingAverage3(speed_list)).

It would be nice to have tcx and gpx parsers yield a sequence of geopoints,
times and other values then have a common library calculate the speeds etc. I'll
leave that for a future change :-)