Patch from Tom Brown:
Short GPX files can trigger a crash in the models.py with the text "Average
speed can not be higher than maximum speed".

Add a test and fix. The fix reorganizes pygpx.parse_segment a bit to split it
into a section that extracts values from the gpx and a section that calculates
distance, speed etc.